created voter's table
included all CRUD operations in API
