+++ server.js
@@ -1,201 +1,27 @@
-//import express.js
-const e = require("express");
-const express = require("express");
-//import mysql2
-const mysql = require("mysql2");
-const inputCheck = require("./utils/inputCheck");
-//server configuration
+const express = require('express');
+const db = require('./db/connection');
+const apiRoutes = require('./routes/apiRoutes');
+
 const PORT = process.env.PORT || 3001;
-//create new express app
 const app = express();
 
-//parse incoming data
-//recognize incoming data as string/array. extended:false = use qs library
+// Express middleware
 app.use(express.urlencoded({ extended: false }));
-//recognize incoming data as a JSON object
 app.use(express.json());
 
-//Connect to database
-const db = mysql.createConnection(
-    {
-        host: 'localhost',
-        user: 'root',
-        password: 'watermelon',
-        database: 'election'
-    },
-    console.log('Connected to the election database.')
-);
+// Use apiRoutes
+app.use('/api', apiRoutes);
 
-app.get('/api/candidates', (req, res) => {
-    const sql = `SELECT candidates.*, parties.name 
-                AS party_name
-                FROM candidates
-                LEFT JOIN parties ON candidates.party_id = parties.id`;
-    db.query(sql, (err, rows) => {
-        if(err) {
-            res.status(500).json({error: err.message});
-            return;
-        }
-        res.json({
-            message: 'success',
-            data: rows
-        });
-    });
-});
-
-//Select a single candidate
-app.get('/api/candidate/:id', (req, res) => {
-    const sql = `SELECT candidates.*, parties.name 
-                AS party_name
-                FROM candidates 
-                LEFT JOIN parties ON candidates.party_id = parties.id
-                WHERE candidates.id=?`;
-    const params = [req.params.id];
-
-    db.query(sql, params, (err, row) => {
-        if(err) {
-            res.status(400).json({error: err.message});
-            return;
-        }
-        res.json({
-            message: 'success',
-            data: row
-        });
-    });
-});
-
-//Delete a candidate
-app.delete('/api/candidate/:id', (req, res) => {
-    const sql = `DELETE FROM candidates WHERE id=?`;
-    const params = [req.params.id];
-    db.query(sql, params, (err, result) => {
-        if(err) {
-            res.statusMessage(400).json({error: err.message});
-        } else if (!result.affectedRows){
-            res.json({
-                message: 'Candidate not found'
-            });
-        } else {
-            res.json({
-                message: 'deleted',
-                changes: result.affectedRows,
-                id: req.params.id
-            });
-        };
-    });
-});
-
-//Create a candidate
-app.post('/api/candidate', ({body}, res) => {
-    const errors = inputCheck(body, 'first_name', 'last_name', 'industry_connected');
-    if(errors) {
-        res.status(400).json({error: err.message});
-        return;
-    };
-
-    const sql = `INSERT INTO candidates (first_name, last_name, industry_connected)
-                VALUES(?,?,?)`;
-    const params = [body.first_name, body.last_name, body.industry_connected];
-    db.query(sql, params, (err, result) => {
-        if(err) {
-            res.status(400).json({error: err.message});
-            return;
-        }
-        res.json({
-            message: 'success',
-            data: body
-        });
-    });
-});
-
-//Update a candidate's party
-app.put('/api/candidate/:id', (req, res) => {
-    //validate request input
-    const errors = inputCheck(req.body, 'party_id');
-    if(errors) {
-        res.status(400).json({error: errors});
-    };
-
-    const sql = `UPDATE candidates SET party_id = ?
-                WHERE id = ?`;
-    const params = [req.body.party_id, req.params.id];
-    db.query(sql, params, (err, result) => {
-        if(err) {
-            res.status(400).json({error: err.message});
-            return;
-        //check if a record was found
-        } else if (!result.affectedRows) {
-            res.json({
-                message: 'Candidate not found'
-            });
-        } else {
-            res.json({
-                message: 'success',
-                data: req.body,
-                changes: result.affectedRows
-            });
-        };
-    });
-});
-
-//Display all parties
-app.get('/api/parties', (req, res) => {
-    const sql = `SELECT * FROM parties`;
-    db.query(sql, (err, rows) => {
-        if(err) {
-            res.status(500).json({error:err.message});
-            return;
-        }
-        res.json({
-            message: 'success',
-            data: rows
-        });
-    });
-});
-
-//Display one party
-app.get('/api/party/:id', (req, res) => {
-    const sql = `SELECT * FROM parties WHERE id=?`;
-    const params = [req.params.id];
-    db.query(sql, params, (err, row) => {
-        if(err) {
-            res.status(400).json({error:err.message});
-            return;
-        } res.json({
-            message: 'success',
-            data: row
-        });
-    });
-});
-
-//Delete a party
-app.delete('/api/party/:id', (req, res) => {
-    const sql = `DELETE FROM parties WHERE id=?`;
-    const params = [req.params.id];
-    db.query(sql, params, (err, result) => {
-        if(err) {
-            res.status(400).json({error:res.message});
-            return;
-        // check if anything was deleted
-        } else if(!result.affectedRows) {
-            res.json({
-                message: 'Party not found'
-            });
-        } else {
-            res.json({
-                message: 'deleted',
-                changes: result.affectedRows,
-                id: req.params.id
-            });
-        };
-    });
-});
-
-//Default response for any other requests (Not Found)
+// Default response for any other request (Not Found)
 app.use((req, res) => {
-    res.status(404).end();
+  res.status(404).end();
 });
 
-app.listen(PORT, () => {
-    console.log(`Server is running at: http://localhost:${PORT}/`);
+// Start server after DB connection
+db.connect(err => {
+  if (err) throw err;
+  console.log('Database connected.');
+  app.listen(PORT, () => {
+    console.log(`Server running on port ${PORT}`);
+  });
 });
